Title: 합과 평균, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240721/합과 평균/sum-and-avg.cpp
@@ -5,6 +5,6 @@ int main() {
     // 여기에 코드를 작성해주세요.
     int a,b;
     cin>>a>>b;
-    cout<<a+b<<" "<<((double(a)+double(b))/2);
+    cout<<a+b<<" "<<double((double(a)+double(b))/2);
     return 0;
 }
